chore: ignore flake8 style warnings
