+++ .flake8
@@ -1,2 +1,3 @@
 [flake8]
 max-line-length = 100
+extend-ignore = E128,E301,E302,E303,E305,E402,E501,E722,F401,F541,F811,F841,W291,W292,W293,W391
