--- app/tests/test_file_processor.py
@@ -1,12 +1,0 @@
-from app.file_processor import detect_file_type
-
-
-def test_detect_file_type() -> None:
-    assert detect_file_type("document.pdf") == "pdf"
-    assert detect_file_type("file.docx") == "docx"
-    assert detect_file_type("script.py") == "py"
-    assert detect_file_type("archive.zip") == "zip"
-    assert detect_file_type("index.html") == "code"
-    assert detect_file_type("notes.txt") == "txt"
-    assert detect_file_type("sheet.xlsx") == "excel"
-    assert detect_file_type("book.xls") == "excel"
