+++ pytest.ini
@@ -1,2 +1,4 @@
 [pytest]
 addopts = -vv
+testpaths = tests
+python_files = test.py
